--- twig_server/neo4j_orm_lite/__init__.py
@@ -1,2 +1,0 @@
-"""Very lite Neo4j 'ORM' designed specifically with this project in mind"""
-from twig_server.neo4j_orm_lite.orm import Twig4jOrm
